+++ src/Tasks/Tasks.WpfUi/App.xaml.cs
@@ -15,128 +15,127 @@ using Tasks.WpfUi.Services;
 using Wpf.Ui.Mvvm.Contracts;
 using Wpf.Ui.Mvvm.Services;
 
-namespace Tasks.WpfUi
+namespace Tasks.WpfUi;
+
+/// <summary>
+/// Interaction logic for App.xaml
+/// </summary>
+public partial class App
 {
+    // The.NET Generic Host provides dependency injection, configuration, logging, and other services.
+    // https://docs.microsoft.com/dotnet/core/extensions/generic-host
+    // https://docs.microsoft.com/dotnet/core/extensions/dependency-injection
+    // https://docs.microsoft.com/dotnet/core/extensions/configuration
+    // https://docs.microsoft.com/dotnet/core/extensions/logging
+    private static readonly IHost _host = Host
+        .CreateDefaultBuilder()
+        .ConfigureAppConfiguration(c => { c.SetBasePath(Path.GetDirectoryName(Assembly.GetEntryAssembly()!.Location)); })
+        .ConfigureServices((context, services) =>
+        {
+            // App Host
+            services.AddHostedService<ApplicationHostService>();
+
+            // Page resolver service
+            services.AddSingleton<IPageService, PageService>();
+
+            // Theme manipulation
+            services.AddSingleton<IThemeService, ThemeService>();
+
+            // TaskBar manipulation
+            services.AddSingleton<ITaskBarService, TaskBarService>();
+
+            // Service containing navigation, same as INavigationWindow... but without window
+            services.AddSingleton<INavigationService, NavigationService>();
+
+            services.AddSingleton<ISnackbarService, SnackbarService>();
+
+            // Main window container with navigation
+            services.AddScoped<INavigationWindow, Views.Container>();
+            services.AddScoped<ViewModels.ContainerViewModel>();
+
+            // Views and ViewModels
+            services.AddScoped<Views.Pages.DashboardPage>();
+            services.AddScoped<ViewModels.DashboardViewModel>();
+                            
+            services.AddScoped<Views.Pages.SettingsPage>();
+            services.AddScoped<ViewModels.SettingsViewModel>();
+
+            services.AddScoped<Views.Pages.RecurrencesPage>();
+            services.AddScoped<ViewModels.RecurrencesPageViewModel>();
+
+            services.AddScoped<Views.Pages.LabelsPage>();
+            services.AddScoped<ViewModels.LabelsPageViewModel>();
+
+            services.AddScoped<Views.Pages.EditLabelPage>();
+            services.AddScoped<ViewModels.EditLabelViewModel>();
+
+            services.AddScoped<Views.Pages.ViewEventPage>();
+            services.AddScoped<ViewModels.ViewEventPageViewModel>();
+
+            services.AddScoped<Views.Pages.AssignedEventLabelsPage>();
+            services.AddScoped<ViewModels.AssignedEventLabelsViewModel>();
+
+            services.AddSingleton<IConfigs, ConfigurationDev>();
+
+            services.AddScoped<WpfApplicationServices>();
+
+            services.AddScoped<IEventServices, EventServices>()
+                .AddScoped<IRecurrenceServices, RecurrenceServices>()
+                .AddScoped<IEventActionServices, EventActionServices>()
+                .AddScoped<IUserServices, UserServices>()
+                .AddScoped<IUserEmailVerificationServices, UserEmailVerificationServices>()
+                .AddScoped<ILabelServices, LabelServices>()
+                .AddScoped<IEventLabelServices, EventLabelServices>()
+
+                // repositories
+                .AddScoped<IUserRepository, UserRepository>()
+                .AddScoped<IEventRepository, EventRepository>()
+                .AddScoped<IRecurrenceRepository, RecurrenceRepository>()
+                .AddScoped<IEventActionRepository, EventActionRepository>()
+                .AddScoped<IUserEmailVerificationRepository, UserEmailVerificationRepository>()
+                .AddScoped<ILabelRepository, LabelRepository>()
+                .AddScoped<IEventLabelRepository, EventLabelRepository>();
+
+
+
+            // Configuration
+            services.Configure<AppConfig>(context.Configuration.GetSection(nameof(AppConfig)));
+        }).Build();
+
     /// <summary>
-    /// Interaction logic for App.xaml
+    /// Gets registered service.
     /// </summary>
-    public partial class App
+    /// <typeparam name="T">Type of the service to get.</typeparam>
+    /// <returns>Instance of the service or <see langword="null"/>.</returns>
+    public static T GetService<T>()
+        where T : class
     {
-        // The.NET Generic Host provides dependency injection, configuration, logging, and other services.
-        // https://docs.microsoft.com/dotnet/core/extensions/generic-host
-        // https://docs.microsoft.com/dotnet/core/extensions/dependency-injection
-        // https://docs.microsoft.com/dotnet/core/extensions/configuration
-        // https://docs.microsoft.com/dotnet/core/extensions/logging
-        private static readonly IHost _host = Host
-            .CreateDefaultBuilder()
-            .ConfigureAppConfiguration(c => { c.SetBasePath(Path.GetDirectoryName(Assembly.GetEntryAssembly()!.Location)); })
-            .ConfigureServices((context, services) =>
-            {
-                // App Host
-                services.AddHostedService<ApplicationHostService>();
+        return _host.Services.GetService(typeof(T)) as T;
+    }
 
-                // Page resolver service
-                services.AddSingleton<IPageService, PageService>();
+    /// <summary>
+    /// Occurs when the application is loading.
+    /// </summary>
+    private async void OnStartup(object sender, StartupEventArgs e)
+    {
+        await _host.StartAsync();
+    }
 
-                // Theme manipulation
-                services.AddSingleton<IThemeService, ThemeService>();
+    /// <summary>
+    /// Occurs when the application is closing.
+    /// </summary>
+    private async void OnExit(object sender, ExitEventArgs e)
+    {
+        await _host.StopAsync();
 
-                // TaskBar manipulation
-                services.AddSingleton<ITaskBarService, TaskBarService>();
+        _host.Dispose();
+    }
 
-                // Service containing navigation, same as INavigationWindow... but without window
-                services.AddSingleton<INavigationService, NavigationService>();
-
-                services.AddSingleton<ISnackbarService, SnackbarService>();
-
-                // Main window container with navigation
-                services.AddScoped<INavigationWindow, Views.Container>();
-                services.AddScoped<ViewModels.ContainerViewModel>();
-
-                // Views and ViewModels
-                services.AddScoped<Views.Pages.DashboardPage>();
-                services.AddScoped<ViewModels.DashboardViewModel>();
-                                
-                services.AddScoped<Views.Pages.SettingsPage>();
-                services.AddScoped<ViewModels.SettingsViewModel>();
-
-                services.AddScoped<Views.Pages.RecurrencesPage>();
-                services.AddScoped<ViewModels.RecurrencesPageViewModel>();
-
-                services.AddScoped<Views.Pages.LabelsPage>();
-                services.AddScoped<ViewModels.LabelsPageViewModel>();
-
-                services.AddScoped<Views.Pages.EditLabelPage>();
-                services.AddScoped<ViewModels.EditLabelViewModel>();
-
-                services.AddScoped<Views.Pages.ViewEventPage>();
-                services.AddScoped<ViewModels.ViewEventPageViewModel>();
-
-                services.AddScoped<Views.Pages.AssignedEventLabelsPage>();
-                services.AddScoped<ViewModels.AssignedEventLabelsViewModel>();
-
-                services.AddSingleton<IConfigs, ConfigurationDev>();
-
-                services.AddScoped<WpfApplicationServices>();
-
-                services.AddScoped<IEventServices, EventServices>()
-                    .AddScoped<IRecurrenceServices, RecurrenceServices>()
-                    .AddScoped<IEventActionServices, EventActionServices>()
-                    .AddScoped<IUserServices, UserServices>()
-                    .AddScoped<IUserEmailVerificationServices, UserEmailVerificationServices>()
-                    .AddScoped<ILabelServices, LabelServices>()
-                    .AddScoped<IEventLabelServices, EventLabelServices>()
-
-                    // repositories
-                    .AddScoped<IUserRepository, UserRepository>()
-                    .AddScoped<IEventRepository, EventRepository>()
-                    .AddScoped<IRecurrenceRepository, RecurrenceRepository>()
-                    .AddScoped<IEventActionRepository, EventActionRepository>()
-                    .AddScoped<IUserEmailVerificationRepository, UserEmailVerificationRepository>()
-                    .AddScoped<ILabelRepository, LabelRepository>()
-                    .AddScoped<IEventLabelRepository, EventLabelRepository>();
-
-
-
-                // Configuration
-                services.Configure<AppConfig>(context.Configuration.GetSection(nameof(AppConfig)));
-            }).Build();
-
-        /// <summary>
-        /// Gets registered service.
-        /// </summary>
-        /// <typeparam name="T">Type of the service to get.</typeparam>
-        /// <returns>Instance of the service or <see langword="null"/>.</returns>
-        public static T GetService<T>()
-            where T : class
-        {
-            return _host.Services.GetService(typeof(T)) as T;
-        }
-
-        /// <summary>
-        /// Occurs when the application is loading.
-        /// </summary>
-        private async void OnStartup(object sender, StartupEventArgs e)
-        {
-            await _host.StartAsync();
-        }
-
-        /// <summary>
-        /// Occurs when the application is closing.
-        /// </summary>
-        private async void OnExit(object sender, ExitEventArgs e)
-        {
-            await _host.StopAsync();
-
-            _host.Dispose();
-        }
-
-        /// <summary>
-        /// Occurs when an exception is thrown by an application but not handled.
-        /// </summary>
-        private void OnDispatcherUnhandledException(object sender, DispatcherUnhandledExceptionEventArgs e)
-        {
-            // For more info see https://docs.microsoft.com/en-us/dotnet/api/system.windows.application.dispatcherunhandledexception?view=windowsdesktop-6.0
-        }
+    /// <summary>
+    /// Occurs when an exception is thrown by an application but not handled.
+    /// </summary>
+    private void OnDispatcherUnhandledException(object sender, DispatcherUnhandledExceptionEventArgs e)
+    {
+        // For more info see https://docs.microsoft.com/en-us/dotnet/api/system.windows.application.dispatcherunhandledexception?view=windowsdesktop-6.0
     }
 }
