+++ src/Tasks/Tasks.WpfUi/CustomConverters/BoolToVisibilityInverseConverter.cs
@@ -4,19 +4,23 @@ using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Data;
 
-namespace Tasks.WpfUi.Helpers;
+namespace Tasks.WpfUi.CustomConverters;
 
-[ValueConversion(typeof(bool), typeof(bool))]
-public class InverseBooleanConverter : IValueConverter
+public class BoolToVisibilityInverseConverter : IValueConverter
 {
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
-        if (targetType != typeof(bool))
-            throw new InvalidOperationException("The target must be a boolean");
+        var booleanVal = (bool)value;
 
-        return !(bool)value;
+        if (booleanVal)
+        {
+            return Visibility.Collapsed;
+        }
+
+        return Visibility.Visible;
     }
 
     public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
