+++ src/Tasks/Tasks.WpfUi/CustomConverters/InverseBooleanConverter.cs
@@ -4,23 +4,19 @@ using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.Windows;
 using System.Windows.Data;
 
-namespace Tasks.WpfUi.Helpers;
+namespace Tasks.WpfUi.CustomConverters;
 
-public class BoolToVisibilityConverter : IValueConverter
+[ValueConversion(typeof(bool), typeof(bool))]
+public class InverseBooleanConverter : IValueConverter
 {
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
-        var booleanVal = (bool)value;
+        if (targetType != typeof(bool))
+            throw new InvalidOperationException("The target must be a boolean");
 
-        if (booleanVal)
-        {
-            return Visibility.Visible;
-        }
-
-        return Visibility.Collapsed;
+        return !(bool)value;
     }
 
     public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
